Messy - Fully Implemented Signals Module Added
Missing tests for the signal class and most documentation, but it
compiles! This is also a test of using multiple branches for
development of features that then go into a develop branch which
then gets pushed as a release to the master branch. This signals
module contains a complete(minus a few depricated features)
implementation of the boost::signals2 library, this includes the
boost::optional library as well. The goal is to have no dependencies
beyond the terminfo/ncurses library.