Add a template parameter to the Layout class.
This is a major change for user code. Any classes that
inherit from a Layout type will need to add `<>` at the
end of the type for default behaviour. Otherwise that
template parameter can be used to limit the type of Widget
that can be used within that layout. make_child() type
methods will default to creating the given Child_t of the
Layout if no template parameter is given.

Also contains various changes in API for accessing child
Widgets. Widget class now has read-only access to children,
while Layout can modify the Children structure.